refactor: Renamed dto to dataclasses
